fix(test): isolate authenticated search canary
